% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_building_code}
\alias{build_building_code}
\title{Build Missouri building code from county-district and building number}
\usage{
build_building_code(county_district_code, building_number)
}
\arguments{
\item{county_district_code}{The 6-digit county-district code}

\item{building_number}{The 4-digit building number}
}
\value{
10-digit building code
}
\description{
Build Missouri building code from county-district and building number
}
\keyword{internal}
