% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{create_empty_building_df}
\alias{create_empty_building_df}
\title{Create empty building data frame with expected columns}
\usage{
create_empty_building_df()
}
\value{
Empty data frame with expected column structure
}
\description{
Create empty building data frame with expected columns
}
\keyword{internal}
