% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_building_via_data_download}
\alias{download_building_via_data_download}
\title{Download building data via school directory data download}
\usage{
download_building_via_data_download(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with building data
}
\description{
Alternative download method using the school directory exports.
}
\keyword{internal}
