% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_dese_excel}
\alias{download_dese_excel}
\title{Download Excel file from Missouri DESE}
\usage{
download_dese_excel(url, type, end_year)
}
\arguments{
\item{url}{URL to download from}

\item{type}{Type of data ("building" or "district")}

\item{end_year}{School year end}
}
\value{
Data frame or NULL if download fails
}
\description{
Download Excel file from Missouri DESE
}
\keyword{internal}
