% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_district_via_data_download}
\alias{download_district_via_data_download}
\title{Download district data via data download}
\usage{
download_district_via_data_download(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with district data
}
\description{
Download district data via data download
}
\keyword{internal}
