% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_mcds_building_data_legacy}
\alias{download_mcds_building_data_legacy}
\title{Download building-level data (legacy format, 2006-2017)}
\usage{
download_mcds_building_data_legacy(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with building-level enrollment data
}
\description{
Download building-level data (legacy format, 2006-2017)
}
\keyword{internal}
