% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_mcds_district_data}
\alias{download_mcds_district_data}
\title{Download district-level data from MCDS (current format, 2018+)}
\usage{
download_mcds_district_data(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with district-level enrollment data
}
\description{
Download district-level data from MCDS (current format, 2018+)
}
\keyword{internal}
