% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from Missouri DESE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
List with building and district data frames
}
\description{
Downloads building and district enrollment data from Missouri DESE's
MCDS system via SSRS report exports.
}
\keyword{internal}
