% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_building_enr}
\alias{process_building_enr}
\title{Process building-level enrollment data}
\usage{
process_building_enr(df, end_year)
}
\arguments{
\item{df}{Raw building data frame}

\item{end_year}{School year end}
}
\value{
Processed building data frame
}
\description{
Process building-level enrollment data
}
\keyword{internal}
